Restructure Project File Organization User/Store Separation
