+++ src/app/create-store/page.tsx
@@ -1,9 +1,9 @@
-import CreateStore from "@/components/store/create-store-form";
+// import CreateStore from "@/components/store/create-store-form";
 
-export default function CreateStorePage() {
-  return (
-    <>
-      <CreateStore />
-    </>
-  );
-}
+// export default function CreateStorePage() {
+//   return (
+//     <>
+//       <CreateStore />
+//     </>
+//   );
+// }
