+++ src/app/page.tsx
@@ -1,11 +1,9 @@
 import LandingPageComponent from "@/components/landing-page/LandingComponent";
-import RecentOrders from "@/components/dashboard/components/RecentOrder";
 
 export default function LadingPage() {
   return (
     <>
       <LandingPageComponent />
-     
     </>
   );
 }
