+++ src/app/register-user/page.tsx
@@ -1,4 +1,4 @@
-import CreateUserForm from "@/components/user/create-user-form"
+import CreateUserForm from "@/components/features/user/create-user-form"
 
 export default function RegisterUserPage() {
     return (
