--- src/app/user-profile/page.tsx
@@ -1,5 +1,0 @@
-import UserProfile from '@/components/user/profile/UserProfile';
-
-export default function UserProfilePage() {
-  return <UserProfile />;
-}
