+++ src/app/user/billing/page.tsx
@@ -1,7 +1,7 @@
 // src/app/billing/page.tsx
 "use client";
 
-import ConnectedWallet from "@/components/billing/connected-wallet";
+import ConnectedWallet from "@/components/features/user/billing/connected-wallet";
 import React from "react";
 
 export default function Page() {
