+++ src/app/user/messages/page.tsx
@@ -1,16 +1,16 @@
 "use client";
 
-import ChatList from "@/components/messages/MessagesList";
-import ChatWindow from "@/components/messages/MessagesWindow";
+import ChatConversations from "@/components/features/user/messages/MessagesList";
+import ChatInterface from "@/components/features/user/messages/MessagesWindow";
 
 export default function ChatLayout() {
   return (
     <div className="flex flex-col md:flex-row w-full min-h-screen items-center justify-center  p-4">
       <div className="w-full md:w-[350px] h-[600px] rounded-2xl p-3">
-        <ChatList />
+        <ChatConversations />
       </div>
       <div className="w-full md:w-[700px] h-[600px] rounded-2xl p-3">
-        <ChatWindow />
+        <ChatInterface />
       </div>
     </div>
   );
