+++ src/app/user/transactions/page.tsx
@@ -1,5 +1,5 @@
 "use client"
-import Transactions from "@/components/transactions/Transaction";
+import Transactions from "@/components/features/user/transactions/Transaction";
 
 export default function TransactionsPage() {
   return (
