+++ src/app/user/user-profile/page.tsx
@@ -1,0 +1,5 @@
+import UserProfile from '@/components/features/user/profile/UserProfile';
+
+export default function UserProfilePage() {
+  return <UserProfile />;
+}
