--- src/components/dashboard/Dashboard.tsx
@@ -1,9 +1,0 @@
-import StoreOverViewComponent from "@/components/dashboard/store-overview/StoreOverViewComponent";
-
-export default function DashboardPage() {
-  return (
-    <div>
-      <StoreOverViewComponent />
-    </div>
-  );
-}
