+++ src/components/features/store/dashboard/Dashboard.tsx
@@ -1,0 +1,9 @@
+import StoreOverview from "./store-overview/StoreOverview";
+
+export default function DashboardPage() {
+  return (
+    <div>
+      <StoreOverview />
+    </div>
+  );
+}
